// src/common/enums/SwaggerTag.ts
export enum SwaggerTag {
    AUTHENTICATION = 'Authentication',
    USER = 'User',
    PROJECT = 'Project',
    EPISODE = 'Episode',
    EPISODE_INSIGHT = 'Episode Insight',
    HEALTH = 'Health',
    STATS = 'Stats',
    ORGANIZATION = 'Organization',
    USER_MANAGEMENT = 'User Management',
    SHARED_EPISODE = 'Shared Episode',
}