// src/common/enums/AuditLogActionName.ts
export const AUDIT_LOG_ACTION_NAME_METADATA_KEY = 'auditLogActionName'

export enum AuditLogActionName {
    // Authentication
    SIGN_UP = 'SIGN_UP',
    CONFIRM_EMAIL = 'CONFIRM_EMAIL',
    SIGN_IN = 'SIGN_IN',
    FORGOT_PASSWORD = 'FORGOT_PASSWORD',
    VALIDATE_FORGOT_PASSWORD_HASH = 'VALIDATE_FORGOT_PASSWORD_HASH',
    SET_NEW_PASSWORD = 'SET_NEW_PASSWORD',
    UPDATE_PASSWORD = 'UPDATE_PASSWORD',

    // User management
    LIST_ORGANIZATION_USERS = 'LIST_ORGANIZATION_USERS',
    ADD_USER_TO_ORGANIZATION = 'ADD_USER_TO_ORGANIZATION',
    SET_USER_ACTIVE_STATUS = 'SET_USER_ACTIVE_STATUS',
    SET_USER_ROLES = 'SET_USER_ROLES',
    DELETE_USER = 'DELETE_USER',
    CONFIRM_ACCOUNT_AND_SET_PASSWORD = 'CONFIRM_ACCOUNT_AND_SET_PASSWORD',

    // Episodes
    LIST_EPISODES = 'LIST_EPISODES',
    GET_EPISODE_BY_PROJECT_ID = 'GET_EPISODE_BY_PROJECT_ID',
    GENERATE_PRE_SIGNED_URL = 'GENERATE_PRE_SIGNED_URL',
    GET_EPISODE_BY_ID = 'GET_EPISODE_BY_ID',
    GET_EPISODE_INSIGHTS_BY_EPISODE_ID = 'GET_EPISODE_INSIGHTS_BY_EPISODE_ID',
    UPDATE_INSIGHTS_BY_EPISODE_ID = 'UPDATE_INSIGHTS_BY_EPISODE_ID',
    GENERATE_INSIGHTS_BY_EPISODE_ID = 'GENERATE_INSIGHTS_BY_EPISODE_ID',

    // Organization
    UPDATE_ORGANIZATION_DETAILS = 'UPDATE_ORGANIZATION_DETAILS',

    // Project
    CREATE_PROJECT = 'CREATE_PROJECT',
    UPDATE_PROJECT_DETAILS = 'UPDATE_PROJECT_DETAILS',

    // Shared episode
    LIST_SHARED_EPISODES = 'LIST_SHARED_EPISODES',
    SHARE_EPISODE = 'SHARE_EPISODE',
    REVOKE_SHARED_EPISODE = 'REVOKE_SHARED_EPISODE',
    UPDATE_SHARED_EPISODE = 'UPDATE_SHARED_EPISODE',
    LIST_SHARED_EPISODES_WITH_USER = 'LIST_SHARED_EPISODES_WITH_USER',
    SHARED_EPISODE_GET_EPISODE_BY_ID = 'SHARED_EPISODE_GET_EPISODE_BY_ID',
    SHARED_EPISODE_GET_EPISODE_INSIGHTS_BY_EPISODE_ID = 'SHARED_EPISODE_GET_EPISODE_INSIGHTS_BY_EPISODE_ID',
    SHARED_EPISODE_GENERATE_INSIGHTS_BY_EPISODE_ID = 'SHARED_EPISODE_GENERATE_INSIGHTS_BY_EPISODE_ID',
    SHARED_EPISODE_UPDATE_INSIGHTS_BY_EPISODE_ID = 'SHARED_EPISODE_UPDATE_INSIGHTS_BY_EPISODE_ID',
}